add Constraints to TaskDef minor refactoring re-added hostmode docker-compose.yml (as a different filename)
