+++ docker-compose-hostmode.yml
@@ -1,0 +1,89 @@
+version: "3"
+
+services:
+  zk:
+    image: zookeeper:3.4
+    ports:
+      - 2188:2188
+    environment:
+      ZOO_MY_ID: 1
+      ZOO_PORT: 2188
+
+  master:
+    image: mesosphere/mesos-master:1.2.1
+    network_mode: host
+    pid: host
+    userns_mode: host
+    privileged: true
+    environment:
+      MESOS_ZK: zk://${DOCKER_IP}:2188/mesos
+      MESOS_QUORUM: 1
+      MESOS_CLUSTER: docker-compose
+      MESOS_HOSTNAME: ${DOCKER_IP}
+    depends_on:
+      - zk
+
+  slave-one:
+    image: mesosphere/mesos-slave:1.2.1
+    network_mode: host
+    pid: host
+    userns_mode: host
+    privileged: true
+    environment:
+      MESOS_MASTER: zk://${DOCKER_IP}:2188/mesos
+      MESOS_CONTAINERIZERS: docker
+      MESOS_PORT: 5051
+      MESOS_RESOURCES: ports(*):[11000-11199]
+      MESOS_HOSTNAME: ${DOCKER_IP}
+      MESOS_ADVERTISE_IP: ${DOCKER_IP}
+      MESOS_WORK_DIR: /tmp/mesos
+      MESOS_ATTRIBUTES: att1:att1valueslave1;att2:att2valueslave1
+    volumes:
+      - /sys/fs/cgroup:/sys/fs/cgroup
+      - /usr/local/bin/docker:/usr/bin/docker
+      - /var/run/docker.sock:/var/run/docker.sock
+    depends_on:
+      - zk
+
+  slave-two:
+    image: mesosphere/mesos-slave:1.2.1
+    network_mode: host
+    pid: host
+    userns_mode: host
+    privileged: true
+    environment:
+      MESOS_MASTER: zk://${DOCKER_IP}:2188/mesos
+      MESOS_CONTAINERIZERS: docker
+      MESOS_PORT: 5052
+      MESOS_RESOURCES: ports(*):[11200-11499]
+      MESOS_HOSTNAME: ${DOCKER_IP}
+      MESOS_ADVERTISE_IP: ${DOCKER_IP}
+      MESOS_WORK_DIR: /tmp/mesos
+      MESOS_ATTRIBUTES: att1:att1valueslave2;att2:att2valueslave2
+    volumes:
+      - /sys/fs/cgroup:/sys/fs/cgroup
+      - /usr/local/bin/docker:/usr/bin/docker
+      - /var/run/docker.sock:/var/run/docker.sock
+    depends_on:
+      - zk
+
+  marathon:
+    image: mesosphere/marathon:v1.3.10
+    entrypoint:
+      - ./bin/start
+      - --disable_ha
+      - --http_port
+      - "8080"
+    network_mode: host
+    pid: host
+    userns_mode: host
+    privileged: true
+    depends_on:
+      - zk
+      - master
+      - slave-one
+      - slave-two
+    environment:
+      - MARATHON_ZK=zk://${DOCKER_IP}:2188/marathon
+      - MARATHON_MASTER=zk://${DOCKER_IP}:2188/mesos
+      - LIBPROCESS_IP=${DOCKER_IP}
