+++ src/main/scala/com/adobe/api/platform/runtime/mesos/TaskBuilder.scala
@@ -15,12 +15,106 @@
 package com.adobe.api.platform.runtime.mesos
 
 import akka.event.LoggingAdapter
+import org.apache.mesos.v1.Protos
+import org.apache.mesos.v1.Protos.CommandInfo
+import org.apache.mesos.v1.Protos.ContainerInfo
+import org.apache.mesos.v1.Protos.ContainerInfo.DockerInfo
 import org.apache.mesos.v1.Protos.ContainerInfo.DockerInfo.PortMapping
+import org.apache.mesos.v1.Protos.HealthCheck
+import org.apache.mesos.v1.Protos.HealthCheck.TCPCheckInfo
+import org.apache.mesos.v1.Protos.NetworkInfo
 import org.apache.mesos.v1.Protos.Offer
+import org.apache.mesos.v1.Protos.Parameter
 import org.apache.mesos.v1.Protos.Resource
+import org.apache.mesos.v1.Protos.TaskID
 import org.apache.mesos.v1.Protos.TaskInfo
+import scala.collection.JavaConverters._
 
 trait TaskBuilder {
   def apply(reqs: TaskDef, offer: Offer, resources: Seq[Resource], portMappings: Seq[PortMapping])(
     implicit logger: LoggingAdapter): TaskInfo
+}
+
+object DefaultTaskBuilder {
+  def apply() = new DefaultTaskBuilder
+}
+class DefaultTaskBuilder extends TaskBuilder {
+
+  def apply(reqs: TaskDef, offer: Offer, resources: Seq[Resource], portMappings: Seq[PortMapping])(
+    implicit logger: LoggingAdapter): TaskInfo = {
+    val healthCheck = reqs.healthCheckPortIndex.map(
+      i =>
+        HealthCheck
+          .newBuilder()
+          .setType(HealthCheck.Type.TCP)
+          .setTcp(TCPCheckInfo
+            .newBuilder()
+            .setPort(reqs.ports(i)))
+          .setDelaySeconds(0)
+          .setIntervalSeconds(1)
+          .setTimeoutSeconds(1)
+          .setGracePeriodSeconds(25)
+          .build())
+
+    val environmentVars = reqs.environment
+      .map(
+        e =>
+          Protos.Environment.Variable.newBuilder
+            .setName(e._1)
+            .setValue(e._2)
+            .build())
+      .asJava
+
+    val parameters = reqs.dockerRunParameters.flatMap {
+      case (k, v) =>
+        v.map(pv => Parameter.newBuilder().setKey(k).setValue(pv).build())
+    }.asJava
+
+    val dockerNetwork = reqs.network match {
+      case _: User => DockerInfo.Network.USER
+      case Host    => DockerInfo.Network.HOST
+      case Bridge  => DockerInfo.Network.BRIDGE
+    }
+
+    //for case of user network, create a single NetworkInfo with the name
+    val networkInfos = reqs.network match {
+      case u: User =>
+        Seq[NetworkInfo](
+          NetworkInfo
+            .newBuilder()
+            .setName(u.name)
+            .build()).asJava
+      case _ => Seq[NetworkInfo]().asJava
+    }
+    val taskBuilder = TaskInfo.newBuilder
+      .setName(reqs.taskName)
+      .setTaskId(TaskID.newBuilder
+        .setValue(reqs.taskId))
+      .setAgentId(offer.getAgentId)
+      .setCommand(
+        CommandInfo.newBuilder
+          .setEnvironment(Protos.Environment.newBuilder
+            .addAllVariables(environmentVars)
+            .build())
+          .setShell(false)
+          .build())
+      .setContainer(
+        ContainerInfo.newBuilder
+          .setType(ContainerInfo.Type.DOCKER)
+          .addAllNetworkInfos(networkInfos)
+          .setDocker(
+            DockerInfo.newBuilder
+              .setImage(reqs.dockerImage)
+              .setNetwork(dockerNetwork)
+              .addAllParameters(parameters)
+              .addAllPortMappings(portMappings.asJava)
+              .build())
+          .build())
+      .addAllResources(resources.asJava)
+    healthCheck match {
+      case Some(hc) => taskBuilder.setHealthCheck(hc)
+      case None     => //no health check
+    }
+    taskBuilder.build()
+  }
 }
